feat: refactored ts files by moving them into lib folder, fixed missing imports and errors when building
