+++ lib/words/words.ts
